Enable dynamic refreshing of client issue mapping from Jira
Before, we needed to manually maintain a list of all Jira issues from our board. Any time we needed to add a new customer, we also had to update our local mapping file.

This commit makes it so we can now retrieve and cache that mapping data for 12 hours. Now, if we have a new customer, the file will be updated (assuming it's expired) and we'll only have to add a nickname to `./config/nicknames.json` if we have one.