Added method to get the Jira issue ID from a client mapping
Adds a method to get the Jira issue ID from a client mapping string. We were doing this manually in `send_worklog()` before. This meant that if you were doing a dry run and had a worklog mismatch, it would have falsely reported that the worklog would have been logged. Breaking this out into its own method, means that the time tracking service classes can use it to check for a match during dry runs.